+++ frontend/src/components/playground/DefaultTransformationContent.tsx
@@ -1,0 +1,9 @@
+export const defaultContent: string = `function main(context) {
+  return (document) => {
+    // Your transformation logic here
+    return document;
+  };
+}
+// Entrypoint function
+(() => main)();
+`;
