+++ frontend/src/components/playground/TransformationBoardItemStrings.tsx
@@ -1,5 +1,4 @@
-"use client";
-export const boardItemI18nStrings = {
+export const transformationBoardItemStrings = {
   dragHandleAriaLabel: "Drag handle",
   dragHandleAriaDescription:
     "Use Space or Enter to activate drag, arrow keys to move, Space or Enter to submit, or Escape to discard. Be sure to temporarily disable any screen reader navigation feature that may interfere with the functionality of the arrow keys.",
